Bug 9117: allocator performance improvements
Fix for following optimization:
Make PodAllocator aware of host tags.

If the service offering has host_tag specified, PodAllocator can pick the pods having hosts with that tag. This would reduce the list of pods the allocator
has to iterate through to find a suitable pod.

Added a new PodAllocator 'HostTagBasedPodAllocator' that extends the 'UserConcentratedAllocator'.