Support SSL/TLS for management server/agent communication
We need to do handshake before any real data transmitted. So we do it in
blocking mode to simplify the process.

Notice we also add a self-signed certificate in utils, which would used as the
certificate for all SSL connection in all mgmt servers. And client won't try to
verify server's certificate. We don't authenicate client as well.