bug 4387: Add maximum limit to the packet, prevent OOME
(Cherry-picked from master tree, for zonesfeature)

The OOME is due to when server reading the data, it would try to adjust the
reading buffer size according to the "packet length" it read. But if the "packet
length" is some random numbers, server would still try to allocate a part of
very big memory for the reading buffer, result in OOM.

This patch add a 64k limit to sending/receiving the packet. It's the maximum
length one IP datagram can support, and we don't think the request can exceed
this limit. Even if exceed the limit in normal condition, we would aware of it
due to the exception.

Solution has been verified using wget and telnet.

status 4387: resolved fixed