Generate self-signed SSL certificate
bug 9459

We would generate SSL certificate when management server start up if we don't
find any in the database and in the local disk(/etc/cloud/management/keystore).
If we got one in database but not in the local disk, we would export the
keystore to local disk. If we got one in local disk, we would update the
database according to the local file.

By this way, we removed the previous pre-installed nioserver.keystore.

status 9459: resolved fix