bug 12586: Remove the duplicate sync code of fullHostSync()
fullHostSync() is messed up.

The sync operation have been done twice, and for the second time, the original
info has already cleared, so default value "Stopped" is created, then result in
stop all the vms.

The fullHostSync() is skipped for Xen, so Xen don't have this issue.

status 12586: resolved fixed