UI: Add quick view to table list
For any list view that has a corresponding detail view, adds a tooltip
to display the compact overview of the detail view contents, including
all available actions. This avoids a user from having to click into a
new panel to see more details and actions for an item.

This will happen automatically for any list view with a 'detailView'
sub-option, and will append a new column to the end of each item
row. If 'noCompact: true' is specified in the detailView's options,
then the quick view will not be rendered.

** Note: This also removes the existing list actions for the instances
         and zones tables.

Example, to show a quick view:

listView: {
  detailView: {
    // Specify noCompact: true to not render a quick view
    //
    // noCompact: true
    ...
  },
  ...
}